src: Removes a lot of responsibility from cli.js Was created two new concepts: * project * license-dependency
Each one will handle things specific from project
and license+dependency.

The work is not done yet, but it will help a lot
with tests in a soon future.
Connects to #211
