+++ lib/file-writer.js
@@ -138,23 +138,6 @@ function copyLicenseFiles (dependencyLicenseFiles) {
   }
 }
 
-// Checks if node_modules exists and returns
-// true if it is not empty not considering '.bin'
-// directory.
-const hasNodeModules = (dir) => {
-  const modulesDir = path.join(dir, 'node_modules');
-  if (fs.existsSync(modulesDir)) {
-    const content = fs.readdirSync(modulesDir).filter(e => e !== '.bin');
-    if (content.length > 0) {
-      return true;
-    } else {
-      return false;
-    }
-  } else {
-    return false;
-  }
-};
-
 module.exports = {
   createXml,
   createHtml,
@@ -162,6 +145,5 @@ module.exports = {
   createJson,
   createYaml,
   createLicenseDir,
-  copyLicenseFiles,
-  hasNodeModules
+  copyLicenseFiles
 };
